Project Package Cleanup
- Moved errors and converters to a separate file within the package
