+++ cla-backend-go/v2/project/errors.go
@@ -1,0 +1,36 @@
+// Copyright The Linux Foundation and each contributor to CommunityBridge.
+// SPDX-License-Identifier: MIT
+
+package project
+
+import (
+	"errors"
+
+	"github.com/communitybridge/easycla/cla-backend-go/gen/v2/models"
+)
+
+// errors
+var (
+	ErrCLAGroupDoesNotExist = errors.New("cla group does not exist")
+	ErrCLAGroupIDMissing    = errors.New("cla group id is missing")
+)
+
+// codedResponse interface
+type codedResponse interface {
+	Code() string
+}
+
+// errorResponse is a helper to wrap the specified error into an error response model
+func errorResponse(err error) *models.ErrorResponse {
+	code := ""
+	if e, ok := err.(codedResponse); ok {
+		code = e.Code()
+	}
+
+	e := models.ErrorResponse{
+		Code:    code,
+		Message: err.Error(),
+	}
+
+	return &e
+}
